refactor: decouple package svgo
