+++ src/svgo/__mocks__/__common__.ts
@@ -1,8 +1,14 @@
-const optimize = jest.fn()
-  .mockReturnValue({ data: "<svg></svg>" })
-  .mockName("SVGOptimizer.optimize");
+const optimizedSvgData = {
+  data: "<svg></svg>",
+};
 
-const optimizer = { optimize };
+const optimize = jest.fn()
+  .mockReturnValue(optimizedSvgData)
+  .mockName("svgo.SVGOptimizer.optimize");
+
+const optimizer = {
+  optimize,
+};
 
 export {
   optimize,
