+++ src/svgo/v1/types.d.ts
@@ -1,7 +1,10 @@
+import type { error, SVGOptimizer } from "../types";
 import type svgo from "svgo-v1";
 
 type SVGOptions = svgo.Options;
 
 export type {
+  error,
+  SVGOptimizer,
   SVGOptions,
 };
