+++ src/svgo/v1/wrapper.ts
@@ -1,5 +1,4 @@
-import type { error } from "../../types";
-import type { SVGOptions } from "./types";
+import type { error, SVGOptions } from "./types";
 import type svgo from "svgo-v1";
 
 import errors from "../../errors";
