+++ src/svgo/v2/types.d.ts
@@ -1,7 +1,10 @@
+import type { error, SVGOptimizer } from "../types";
 import type svgo from "svgo-v2"; // eslint-disable-line import/default
 
 type SVGOptions = svgo.Options;
 
 export type {
+  error,
+  SVGOptimizer,
   SVGOptions,
 };
