+++ src/svgo/v2/wrapper.ts
@@ -1,5 +1,4 @@
-import type { error } from "../../types";
-import type { SVGOptions } from "./types";
+import type { error, SVGOptions } from "./types";
 import type svgo from "svgo-v2"; // eslint-disable-line import/default
 
 import errors from "../../errors";
